requirements.txt
pandas
langchain
langchain-google-genai
langchain-community
langchain-core
faiss-cpu
python-dotenv
fastapi
uvicorn